update code 20180129

+++ src/ui/api/member.go
@@ -1,33 +1,34 @@
-// Copyright (c) 2017 VMware, Inc. All Rights Reserved.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//    http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
+/*
+   Copyright (c) 2016 VMware, Inc. All Rights Reserved.
+   Licensed under the Apache License, Version 2.0 (the "License");
+   you may not use this file except in compliance with the License.
+   You may obtain a copy of the License at
+
+       http://www.apache.org/licenses/LICENSE-2.0
+
+   Unless required by applicable law or agreed to in writing, software
+   distributed under the License is distributed on an "AS IS" BASIS,
+   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+   See the License for the specific language governing permissions and
+   limitations under the License.
+*/
 
 package api
 
 import (
 	"fmt"
 	"net/http"
-	"strings"
+	"strconv"
 
+	"github.com/vmware/harbor/src/common/api"
 	"github.com/vmware/harbor/src/common/dao"
 	"github.com/vmware/harbor/src/common/models"
 	"github.com/vmware/harbor/src/common/utils/log"
-	"github.com/vmware/harbor/src/ui/auth"
 )
 
 // ProjectMemberAPI handles request to /api/projects/{}/members/{}
 type ProjectMemberAPI struct {
-	BaseController
+	api.BaseAPI
 	memberID      int
 	currentUserID int
 	project       *models.Project
@@ -41,83 +42,47 @@ type memberReq struct {
 
 // Prepare validates the URL and parms
 func (pma *ProjectMemberAPI) Prepare() {
-	pma.BaseController.Prepare()
-
-	if !pma.SecurityCtx.IsAuthenticated() {
-		pma.HandleUnauthorized()
-		return
-	}
-	user, err := dao.GetUser(models.User{
-		Username: pma.SecurityCtx.GetUsername(),
-	})
+	pid, err := strconv.ParseInt(pma.Ctx.Input.Param(":pid"), 10, 64)
 	if err != nil {
-		pma.HandleInternalServerError(
-			fmt.Sprintf("failed to get user %s: %v",
-				pma.SecurityCtx.GetUsername(), err))
+		log.Errorf("Error parsing project id: %d, error: %v", pid, err)
+		pma.CustomAbort(http.StatusBadRequest, "invalid project Id")
 		return
 	}
-	pma.currentUserID = user.UserID
-
-	pid, err := pma.GetInt64FromPath(":pid")
-	if err != nil || pid <= 0 {
-		text := "invalid project ID: "
-		if err != nil {
-			text += err.Error()
-		} else {
-			text += fmt.Sprintf("%d", pid)
-		}
-		pma.HandleBadRequest(text)
-		return
-	}
-	project, err := pma.ProjectMgr.Get(pid)
+	p, err := dao.GetProjectByID(pid)
 	if err != nil {
-		pma.ParseAndHandleError(fmt.Sprintf("failed to get project %d", pid), err)
-		return
-	}
-	if project == nil {
-		pma.HandleNotFound(fmt.Sprintf("project %d not found", pid))
-		return
-	}
-	pma.project = project
-
-	if !(pma.Ctx.Input.IsGet() && pma.SecurityCtx.HasReadPerm(pid) ||
-		pma.SecurityCtx.HasAllPerm(pid)) {
-		pma.HandleForbidden(pma.SecurityCtx.GetUsername())
-		return
+		log.Errorf("Error occurred in GetProjectById, error: %v", err)
+		pma.CustomAbort(http.StatusInternalServerError, "Internal error.")
 	}
 
-	if len(pma.GetStringFromPath(":mid")) != 0 {
-		mid, err := pma.GetInt64FromPath(":mid")
-		if err != nil || mid <= 0 {
-			text := "invalid member ID: "
-			if err != nil {
-				text += err.Error()
-			} else {
-				text += fmt.Sprintf("%d", mid)
-			}
-			pma.HandleBadRequest(text)
-			return
-		}
-
-		member, err := dao.GetUser(models.User{
-			UserID: int(mid),
-		})
+	if p == nil {
+		log.Warningf("Project with id: %d does not exist.", pid)
+		pma.CustomAbort(http.StatusNotFound, "Project does not exist")
+	}
+	pma.project = p
+	pma.currentUserID = pma.ValidateUser()
+	mid := pma.Ctx.Input.Param(":mid")
+	if mid == "current" {
+		pma.memberID = pma.currentUserID
+	} else if len(mid) == 0 {
+		pma.memberID = 0
+	} else if len(mid) > 0 {
+		memberID, err := strconv.Atoi(mid)
 		if err != nil {
-			pma.HandleInternalServerError(fmt.Sprintf("failed to get user %d: %v", mid, err))
-			return
+			log.Errorf("Invalid member Id, error: %v", err)
+			pma.CustomAbort(http.StatusBadRequest, "Invalid member id")
 		}
-		if member == nil {
-			pma.HandleNotFound(fmt.Sprintf("member %d not found", mid))
-			return
-		}
-
-		pma.memberID = member.UserID
+		pma.memberID = memberID
 	}
 }
 
 // Get ...
 func (pma *ProjectMemberAPI) Get() {
 	pid := pma.project.ProjectID
+	if !checkProjectPermission(pma.currentUserID, pid) {
+		log.Warningf("Current user, user id: %d does not have permission for project, id: %d", pma.currentUserID, pid)
+		pma.RenderError(http.StatusForbidden, "")
+		return
+	}
 	if pma.memberID == 0 { //member id not set return list of the members
 		username := pma.GetString("username")
 		queryUser := models.User{Username: username}
@@ -156,42 +121,22 @@ func (pma *ProjectMemberAPI) Get() {
 
 // Post ...
 func (pma *ProjectMemberAPI) Post() {
+	currentUserID := pma.currentUserID
 	projectID := pma.project.ProjectID
+	if !hasProjectAdminRole(currentUserID, projectID) {
+		log.Warningf("Current user, id: %d does not have project admin role for project, id:", currentUserID, projectID)
+		pma.RenderError(http.StatusForbidden, "")
+		return
+	}
 
 	var req memberReq
 	pma.DecodeJSONReq(&req)
-	username := strings.TrimSpace(req.Username)
+	username := req.Username
 	userID := checkUserExists(username)
 	if userID <= 0 {
-
-		user, err := auth.SearchUser(username)
-
-		if err != nil {
-			log.Errorf("Failed the search user, error: %v", err)
-			pma.RenderError(http.StatusInternalServerError, "Failed to search user")
-			return
-		}
-
-		if user == nil {
-			log.Errorf("Current user doesn't exist: %v", username)
-			pma.RenderError(http.StatusNotFound, "Failed to search user: "+username)
-			return
-		}
-
-		err = auth.OnBoardUser(user)
-
-		if err != nil {
-			log.Errorf("Failed the onboard user, error: %s", err)
-			pma.RenderError(http.StatusInternalServerError, "Failed to onboard user")
-			return
-		}
-		if user.UserID <= 0 {
-			log.Error("Failed the onboard user, UserId <=0")
-			pma.RenderError(http.StatusInternalServerError, "Failed to onboard user")
-			return
-		}
-		userID = user.UserID
-
+		log.Warningf("User does not exist, user name: %s", username)
+		pma.RenderError(http.StatusNotFound, "User does not exist")
+		return
 	}
 	rolelist, err := dao.GetUserProjectRoles(userID, projectID)
 	if err != nil {
@@ -225,7 +170,14 @@ func (pma *ProjectMemberAPI) Post() {
 
 // Put ...
 func (pma *ProjectMemberAPI) Put() {
+	currentUserID := pma.currentUserID
 	pid := pma.project.ProjectID
+	if !hasProjectAdminRole(currentUserID, pid) {
+		log.Warningf("Current user, id: %d does not have project admin role for project, id:", currentUserID, pid)
+		pma.RenderError(http.StatusForbidden, "")
+		return
+	}
+
 	mid := pma.memberID
 
 	var req memberReq
@@ -257,7 +209,14 @@ func (pma *ProjectMemberAPI) Put() {
 
 // Delete ...
 func (pma *ProjectMemberAPI) Delete() {
+	currentUserID := pma.currentUserID
 	pid := pma.project.ProjectID
+	if !hasProjectAdminRole(currentUserID, pid) {
+		log.Warningf("Current user, id: %d does not have project admin role for project, id:", currentUserID, pid)
+		pma.RenderError(http.StatusForbidden, "")
+		return
+	}
+
 	mid := pma.memberID
 
 	err := dao.DeleteProjectMember(pid, mid)
